map can scroll

+++ include/type.hpp
@@ -4,7 +4,7 @@
 #include <tuple>
 #include <stdint.h>
 
-typedef float scalar_t;
+typedef int scalar_t;
 typedef int64_t id_type;
 typedef std::tuple<scalar_t, scalar_t, scalar_t> coor_t;
